petcare-connect/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Felicia Onye <johnoluwafemi64@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.67.0"
gino = {extras = ["pg", "starlette"], version = "^1.0.1"}
uvicorn = "^0.21.1"
gunicorn = "^20.1.0"
alembic = "^1.10.2"
psycopg2-binary = "^2.9.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email", "dotenv"], version = "^1.8.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]
pytest = "^5.0"
pytest-cov = "^2.10.1"
pytest-asyncio = "^0.15.1"
httpx = "^0.18.1"

[tool.poetry.group.dev.dependencies]
pytest = "^5.2"
requests = "^2.28.2"
pytest-cov = "^2.10.1"
pytest-asyncio = "^0.15.1"
httpx = "^0.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
